hack main.js, closer to exercising what we care about
Note that is isn't yet correct: we won't really be using Workers in the final
app, because we need synchronous/blocking calls back up to the dispatch unit.
A proper demo will start with a C program that uses XS API calls and "host
functions" to connect the dispatch unit with the snapshotted vat.

Using Workers bypasses the requirement that there be a way to access a
well-known function in the vat that was loaded from a snapshot, as well as
providing host functions to the newly-loaded vat. With Workers, `postMessage`
and `worker.onmessage=` are sufficient.